First rough draft of the physics simulator
Small physics simulator where you can change the initial values in the config file and then can change them during the simulation through slide bars at the bottom. Optimizations need to be done such as for the collisions and also need to make it so that the balls don't slow down when I change the damp value.